+++ packages/expo/src/cache/types.ts
